+++ cnccode/data_classes/DBEJCallActivity.php
@@ -565,4 +565,15 @@ class DBEJCallActivity extends DBECallActivity
         $this->setQueryString($query);
         return (parent::getRows());
     }
+
+    public function getPendingSalesRequestRows()
+    {
+        $query =
+            "SELECT " .
+            $this->getDBColumnNamesAsString() .
+            " FROM " . $this->fromString .
+            " WHERE callactivity.caa_status = 'O' and caa_callacttypeno = 43";
+        $this->setQueryString($query);
+        return (parent::getRows());
+    }
 }