+++ htdocs/config.inc.php
@@ -695,11 +695,14 @@ define(
 define(
     'CONFIG_FIXED_ACTIVITY_TYPE_ID',
     57
-);
-define(
+
+);define(
     'CONFIG_OPERATIONAL_ACTIVITY_TYPE_ID',
     60
 );
+define(
+    'CONFIG_SALES_ACTIVITY_TYPE_ID',
+    43);
 define(
     'CONFIG_SERVER_HEALTH_CHECK_ACTIVITY_TYPE_ID',
     12
@@ -707,20 +710,20 @@ define(
 define(
     'CONFIG_SERVER_HEALTH_CHECK_CHECKLIST_ACTIVITY_TYPE_ID',
     48
-);
-define(
+
+);define(
     'CONFIG_SERVER_HEALTH_CHECK_OFF_SITE_ACTIVITY_TYPE_ID',
     12
-);
-define(
+
+);define(
     'CONFIG_SERVER_HEALTH_CHECK_ON_SITE_ACTIVITY_TYPE_ID',
     50
-);
-define(
+
+);define(
     'CONFIG_SERVER_GUARD_UPDATE_ACTIVITY_TYPE_ID',
     55
-);
-define(
+
+);define(
     'CONFIG_2NDSITE_BACKUP_ACTIVITY_TYPE_ID',
     49
 );